strativ_test/apps/countries/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView

from strativ_test.apps.countries.filters import CountryFilter
from strativ_test.apps.countries.models import Country
from strativ_test.apps.countries.serializers import CountrySerializer


class CountryListAPIView(ListCreateAPIView):
    serializer_class = CountrySerializer
    queryset = Country.objects.all()
    filter_backends = [DjangoFilterBackend]
    filter_class = CountryFilter
    # permission_classes = []


class CountryRetrieveUpdateDestroyAPIView(RetrieveUpdateDestroyAPIView):
    serializer_class = CountrySerializer
    queryset = Country.objects.all()
    filter_backends = [DjangoFilterBackend]
    # permission_classes = []
