requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
python-decouple==3.4
urlwait==1.0
dj-config-url~=0.1.1
psycopg2-binary==2.8.6
requests==2.25.1
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.7.0